New Feature: Thread handling (#6)
Adds the ability to continue previous threads by selecting them from a list.

Updates the DB logic around threads to also save the initial prompt, for referencing while selecting previous threads.

Adds TerminalSelector class that uses curses to display an interactive list of threads to make the selection.

Refactors commands in the IO loop to be selected via a lookup

Adds Command protocol and builds command class (with __call__ implemented) which is used to define each command function.

Each command takes in an environ arg as its only arg, via which it can access properties and manipulate them, meaning the manipulated properties are available to the calling code.